requirements.txt
pillow
pandas
ollama